Two steps are added to the regional iterative tracking (RegIt) for heavy ion muons: the detached triplet step, and the "muon-seeded step", based on the equivalent step from pp iterative tracking.
The effect of this change is to slightly enhance the muon reconstruction efficiency, especially at high pt and in the endcaps, with very small impact on reconstruction timing.
Some illustrative plots can be found in slides 5-6 of https://indico.cern.ch/event/441327/contribution/11/attachments/1144636/1640681/chapon_dilepton_20150826.pdf.