@cmsbuild 2015-08-26
A new Pull Request was created by @echapon (Emilien Chapon) for CMSSW_7_6_X.

New steps for HI muon RegIt

It involves the following packages:

RecoHI/HiMuonAlgos
RecoHI/HiTracking


@cmsbuild, @cvuosalo, @slava77 can you please review it and eventually sign? Thanks.
@miheejo, @jazzitup, @richard-cms, @yenjie, @kurtejung, @istaslis, @mandrenguyen, @dgulhan, @yetkinyilmaz this is something you requested to watch as well.
You can sign-off by replying to this message having '+1' in the first line of your reply.
You can reject by replying  to this message having '-1' in the first line of your reply.
If you are a L2 or a release manager you can ask for tests by saying 'please test' in the first line of a comment.
@degano you are the release manager for this.
You can merge this pull request by typing 'merge' in the first line of your comment.