@cmsbuild 2015-08-26
The tests are being triggered in jenkins.