@cmsbuild 2015-08-26
+1
Tested at: 03edf1fd5989489b3a6d8c5eb0e262291a029939
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-10952/7668/summary.html