@cmsbuild 2015-08-26
Comparison is ready
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-10952/7668/summary.html