@cvuosalo 2015-08-26
@echapon: Is there any relationship between this PR and #10653? Or are they independent?