@echapon 2015-08-26
These two PRs both affect muon regional iterative tracking for heavy ions, but they are independent.