@cvuosalo 2015-08-27
+1

For #10952 03edf1fd5989489b3a6d8c5eb0e262291a029939

For Heavy Ion muons, adding regional iterative tracking and detached triplet and muon-seeded steps.

The code changes are satisfactory, and Jenkins [tests](https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-10952/7668/summary.html) against baseline CMSSW_7_6_X_2015-08-26-1100 show only very tiny differences, which, though numerous, are not significant. A test of workflow 140.53 with 70 events against baseline CMSSW_7_6_X_2015-08-24-1100 also shows a number of tiny differences which are not significant, with two examples shown at bottom. Processing time increases by about 4%:
```
Baseline
Max VSIZ 3145.04 on evt 38 ; max RSS 2697.1 on evt 59
Time av 22.2036 s/evt   max 108.176 s on evt 59
M1 Time av 22.2325 s/evt   max 108.176 s on evt 59

PR
Max VSIZ 3189.58 on evt 38 ; max RSS 2694.92 on evt 59
Time av 23.0261 s/evt   max 113.079 s on evt 59
M1 Time av 23.0663 s/evt   max 113.079 s on evt 59
```
Another measure of the Baseline==>PR timing difference:
```
Job total:  22.0545 s/ev ==> 22.8502 s/ev
Job total:  22.2708 s/ev ==> 23.0769 s/ev (excluding first event)
```
Times for the newly enabled modules (including first event):
```
       added      +0.82%         0.00 ms/ev ->       180.81 ms/ev hiEarlyMuons
       added      +0.00%         0.00 ms/ev ->         1.01 ms/ev hiRegitMuDetachedTripletStepSeedLayers
       added      +0.15%         0.00 ms/ev ->        34.16 ms/ev hiRegitMuDetachedTripletStepTrackCandidates
       added      +0.18%         0.00 ms/ev ->        40.30 ms/ev hiRegitMuonSeededTrackCandidatesInOut
       added      +2.11%         0.00 ms/ev ->       464.92 ms/ev hiEarlyGeneralTracks
       added      +0.04%         0.00 ms/ev ->         8.34 ms/ev hiRegitMuonSeededTracksInOut
       added      +0.02%         0.00 ms/ev ->         4.16 ms/ev hiRegitMuDetachedTripletStepTracks
       added      +0.00%         0.00 ms/ev ->         0.17 ms/ev hiRegitMuonSeededTrackCandidatesOutIn
       added      +0.00%         0.00 ms/ev ->         0.06 ms/ev hiRegitMuonSeededTracksOutInSelector
       added      +0.00%         0.00 ms/ev ->         0.05 ms/ev hiRegitMuDetachedTripletStepClusters
       added      +0.01%         0.00 ms/ev ->         2.47 ms/ev hiRegitMuonSeededSeedsInOut
       added      +0.00%         0.00 ms/ev ->         0.23 ms/ev hiRegitMuonSeededTracksInOutSelector
       added      +0.00%         0.00 ms/ev ->         0.08 ms/ev hiRegitMuonSeededTracksOutIn
       added      +0.02%         0.00 ms/ev ->         4.97 ms/ev hiRegitMuDetachedTripletStepSeeds
       added      +0.00%         0.00 ms/ev ->         0.85 ms/ev hiRegitMuonSeededSeedsOutIn
       added      +0.00%         0.00 ms/ev ->         0.14 ms/ev hiRegitMuDetachedTripletStepSelector
```
Examples of small differences:
![jetptratio](https://cloud.githubusercontent.com/assets/5736159/9511047/dbd2ae9c-4c37-11e5-9397-43bff7220879.png)
![numtracks](https://cloud.githubusercontent.com/assets/5736159/9511048/dbd5b52e-4c37-11e5-8ede-d0ba8f336952.png)